+++ lib/omnyshell_client.dart
@@ -16,6 +16,8 @@ import 'src/application/client/client_runtime.dart';
 export 'omnyshell.dart';
 
 export 'src/application/client/client_runtime.dart';
+export 'src/application/client/command_classifier.dart'
+    show launchesForegroundProgram, mayChangeCwdOrGit;
 export 'src/application/client/cwd_marker.dart';
 export 'src/application/client/file_transfer.dart'
     show downloadPath, uploadPath, ClientTransferLink;